+++ app/index.js
@@ -1,7 +1,13 @@
 import React from 'react';
 import EStyleSheet from 'react-native-extended-stylesheet';
-import {StyleSheet} from 'react-native'
-import Home from './screens/Home'
+import {StyleSheet} from 'react-native';
+
+import {createAppContainer, createStackNavigator} from 'react-navigation';
+
+// screens
+import Home from './screens/Home';
+import Map from './screens/Map';
+
 
 EStyleSheet.build({
     //put global styles here
@@ -9,4 +15,17 @@ EStyleSheet.build({
     $inputText: '#000000',
 });
 
-export default () => <Home/>;
+const App = createAppContainer(
+  // screen navigator
+  createStackNavigator(
+    {
+      Home: {screen: Home},
+      Map: {screen: Map},
+    },
+    {
+      headerMode: 'screen',
+    }
+  )
+);
+
+export default App;
