+++ package.json
@@ -10,7 +10,9 @@
     "react": "16.6.3",
     "react-native": "0.57.8",
     "react-native-extended-stylesheet": "^0.11.1",
-    "react-native-maps": "^0.22.1"
+    "react-native-gesture-handler": "^1.0.12",
+    "react-native-maps": "^0.22.1",
+    "react-navigation": "^3.0.9"
   },
   "devDependencies": {
     "babel-jest": "23.6.0",
